% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{adapt_manual_tree}
\alias{adapt_manual_tree}
\title{adapt_manual_tree}
\usage{
adapt_manual_tree(InputTreeDF)
}
\arguments{
\item{InputTreeDF}{The input tree from the user interface, the interface is defined
as such. columns should be c('left daughter','right daughter','split var','split point','prediction'.
For non terminal nodes, the left and right daughter columns should give the numbers of each node
for terminal nodes, both left and right daughter values should be 0. For nonterminal nods
Split var should include the name of the variable used for splitting at that node. Leave blank for terminal nodes.
the split point should give the point at which to split the node. If it is less than the value
it will flow to left daughtr, if it is greater it will go to right daughter. prediction needs only
 to be filled out for terminal nodes and should indicate the desired prediction.}
}
\value{
The input tree df compatible with the rest of the package
}
\description{
Adapt trees defined from the user interface to the format interfaced with
the rest of the package
}
