% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{adapt_rf2rpart}
\alias{adapt_rf2rpart}
\title{adapt_rf2rpart}
\usage{
adapt_rf2rpart(rf_tree, add_data, X, y)
}
\arguments{
\item{rf_tree}{the random forest tree object}

\item{add_data}{should information from a compatible df
be used to populate tree info? T/F}

\item{X}{if add_data = T, what is the predictor matrix}

\item{y}{if add_data = T  what are the true values?}
}
\description{
adapter function designed to take information from the rf_tree object to the 
rpart style object so that the plotting capabilities of the rpart object can be 
leveraged.
}
