% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{add_zero_columns}
\alias{add_zero_columns}
\title{add_zero_columns
Add Columns with 0s for Each Unique Value in a Specified Column}
\usage{
add_zero_columns(df, column_name)
}
\arguments{
\item{df}{A data.frame to which the new columns will be added.}

\item{column_name}{A character string representing the name of the column in 'df' containing the unique values for creating new columns.}
}
\value{
A data.frame with new columns initialized with 0s for each unique value in the specified column.
}
\description{
This function adds new columns to a dataframe for each unique value in a specified column, initializing these new columns with 0s.
}
\examples{
# Create an example dataframe
data <- data.frame(
  A = c("apple", "banana", "apple", "banana", "orange"),
  B = c(1, 2, 3, 4, 5)
)

# Use the add_zero_columns function with dplyr
library(dplyr)
result <- data \%>\%
  mutate(data, add_zero_columns(., "A"))
print(result)
}
