% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{fill_tree_data}
\alias{fill_tree_data}
\title{fill_tree_data}
\usage{
fill_tree_data(rpart_obj, rf_tree, X, y)
}
\arguments{
\item{rpart_obj}{the rpart object to modify}

\item{rf_tree}{the rf_tree object representing the tree to fill}

\item{X}{the data used to train the model or data of an equivilant format
no true labels}

\item{y}{the true labels for the training data}
}
\value{
an rpart object with the split information correctly filled out based on
the supplied information
}
\description{
fills the tree with information from which it was trained
}
