% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{get_num_from_factor_levels}
\alias{get_num_from_factor_levels}
\title{get_num_from_factor_levels}
\usage{
get_num_from_factor_levels(predicts)
}
\arguments{
\item{rf_tree}{the rf_tree object}
}
\value{
a numeric vector corresponding to the prediction column from rf_tree
}
\description{
Transforms preidctions in whatever form into the numeric version factor levels
within the rpart_frame object
}
