% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{init_rpart}
\alias{init_rpart}
\title{init_rpart}
\usage{
init_rpart(model_type)
}
\arguments{
\item{model_type}{Whether to return a model for 'regression' or a 'categorical' prediction}
}
\value{
an rpart object fit on the iris dataset
}
\description{
This function fit a simple rpart object to the iris dataset to generate an rpart object
which can be modified
}
