% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{init_rpart_frame_matrix}
\alias{init_rpart_frame_matrix}
\title{init_rpart_frame_matrix}
\usage{
init_rpart_frame_matrix(rf_tree)
}
\arguments{
\item{the}{rftree object to get the information from}
}
\description{
initializes the matrix portion of the rpart$frame
}
