% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{iris_test}
\alias{iris_test}
\title{iris_test}
\usage{
iris_test(testdir, test_type = "empty")
}
\arguments{
\item{testdir}{where to save the plots}

\item{test_type}{which iris test to run, options "empty", "filled"}
}
\description{
iris_test
}
