% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{make_rownames_binary_tree}
\alias{make_rownames_binary_tree}
\title{make_rownames_binary_tree}
\usage{
make_rownames_binary_tree(rf_tree)
}
\arguments{
\item{rf_tree}{the rf_tree object}
}
\value{
a vector of numbers of the new rownames
}
\description{
the rownames of rpart$frame need to be in the binary tree format. THis function
calculates those rownames based on the relationships list in the rf_tree object
}
