% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{make_rpart_tests}
\alias{make_rpart_tests}
\title{make_rpart_tests}
\usage{
make_rpart_tests()
}
\value{
named list of tests
}
\description{
Generates models to test using test_rpart, save the results in a named list
}
