% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{mk_confusion_mat}
\alias{mk_confusion_mat}
\title{mk_confusion_matrix}
\usage{
mk_confusion_mat(tree_pref_obj, save_loc = "~/confmatsz.jpeg")
}
\arguments{
\item{confusion_data}{a dataframe with two columns, the first being predictions and
the second being the true values. a compatible object is output by tree_performance in
the $conf_mat attribute}

\item{saveloc}{where should the confusion matrix be stored}
}
\description{
Produces a heatmap that functions as a confusion matrix using output of
tree_performance.
}
