% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{mod_rpart_frame}
\alias{mod_rpart_frame}
\title{mod_rpart_frame}
\usage{
mod_rpart_frame(rpart_obj, input_tree_df)
}
\arguments{
\item{rpart_obj}{the rpart shell to modify}

\item{input_tree_df}{the random forest tree}
}
\value{
modified rpart object
}
\description{
Transfer information from the input format to the frame aspect of the rpart object
}
