% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{mod_rpart_splits}
\alias{mod_rpart_splits}
\title{mod_rpart_splits}
\usage{
mod_rpart_splits(rpart_obj, input_tree_df)
}
\arguments{
\item{rpart_obj}{the rpart object to be modified}

\item{input_tree_df}{the input tree df to get the split information from}
}
\value{
modified rpart_object
}
\description{
modifies the split aspect of the rpart_object dataframe to contain information
from the input_tree_df object
}
