% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{mod_rpart_ylevels}
\alias{mod_rpart_ylevels}
\title{mod_rpart_ylevels}
\usage{
mod_rpart_ylevels(rpart_obj, rf_tree)
}
\arguments{
\item{rpart_obj}{the rpart object to be modified}

\item{rf_tree}{the rf_tree object to get the names from}
}
\value{
modified rpart_object
}
\description{
Supplies new levels to the rpart object based on the predictions from the rf_tree
notably adds a blank string "" to the list to supply to nonterminal nodes as a
placeholder
}
