% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{QckRBrwrPllt}
\alias{QckRBrwrPllt}
\title{QckRBrwrPllt}
\usage{
QckRBrwrPllt(name, n)
}
\arguments{
\item{name}{the Rbrewer pallette name}

\item{n}{how many colors you want}
}
\value{
a color vector
}
\description{
This is a quick function that combines two RColorBrewer functions into one that
makes it more intuitive to use and setup
}
