% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{test_mod_rpart}
\alias{test_mod_rpart}
\title{test_mod_rpart}
\usage{
test_mod_rpart(model_list, save_folder)
}
\arguments{
\item{model_list}{named list of models}

\item{save_folder}{where to save the plotted trees}
}
\value{
plots matching one tree from each model
}
\description{
Take a list of models run them through the main function, and store the results
}
