% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{init_rpart}
\alias{init_rpart}
\title{init_rpart}
\usage{
init_rpart()
}
\value{
an rpart object fit on the iris dataset
}
\description{
This function fit a simple lm to the iris dataset so as to generate an rpart objec
which can then be used as a shell for the rest of the code
}
