% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{init_rpart_frame_matrix}
\alias{init_rpart_frame_matrix}
\title{init_rpart_frame_matrix}
\usage{
init_rpart_frame_matrix(input_tree_df)
}
\arguments{
\item{input_tree_df}{the input df defining the decision tree to get the information from}
}
\description{
initializes the matrix portion of the rpart$frame with the format for the
correct number of classes
}
