% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{make_rownames_binary_tree}
\alias{make_rownames_binary_tree}
\title{make_rownames_binary_tree}
\usage{
make_rownames_binary_tree(input_tree_df)
}
\arguments{
\item{input_tree_df}{the input_tree_df object}
}
\value{
a vector of numbers of the new rownames
}
\description{
The rownames of rpart$frame need to be in the binary tree format. This function
calculates those rownames based on the relationships defined in the input_df object
}
