% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{mod_rpart_frame}
\alias{mod_rpart_frame}
\title{mod_rpart_frame}
\usage{
mod_rpart_frame(rpart_obj, rf_tree)
}
\arguments{
\item{rpart_obj}{the rpart shell to modify}

\item{rf_tree}{the random forest tree}
}
\value{
modified rpart object
}
\description{
this function transfers information from the rf_tree to the frame aspect of the
rpart object
}
