% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analyzer.R
\name{mod_rpart_splits}
\alias{mod_rpart_splits}
\title{mod_rpart_splits}
\usage{
mod_rpart_splits(rpart_obj, rf_tree)
}
\arguments{
\item{rpart_obj}{the rpart object to be modified}

\item{rf_tree}{the rf_tree object to get the split information from}
}
\value{
modified rpart_object
}
\description{
modifies the split aspect of the rpart_object dataframe to contain information
from the rf_tree object
}
